requirements.txt
torch
torchvision
mlflow
Pillow
scikit-learn
numpy
tqdm
fastapi
uvicorn[standard]
python-multipart
pytest
pytest-html
pytest-cov